requirements.txt
Pillow==9.3.0
numpy==1.23.5
torch==2.0.1+cu118
typing==3.7.4.3
opencv-python==4.6.0.66
torchvision==0.15.2+cu118